+++ src/django/api/migrations/0213_add_can_get_union_linked_data_group.py
@@ -1,0 +1,31 @@
+from django.db import migrations
+
+from api.constants import FeatureGroups
+
+
+def create_can_get_union_linked_data(apps, schema_editor):
+    Group = apps.get_model('auth', 'Group')
+    Group.objects.get_or_create(
+        name=FeatureGroups.CAN_GET_UNION_LINKED_DATA
+    )
+
+
+def remove_can_get_union_linked_data(apps, schema_editor):
+    Group = apps.get_model('auth', 'Group')
+    Group.objects.filter(
+        name=FeatureGroups.CAN_GET_UNION_LINKED_DATA
+    ).delete()
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('api', '0212_add_moderation_pause_info_switch'),
+    ]
+
+    operations = [
+        migrations.RunPython(
+            create_can_get_union_linked_data,
+            remove_can_get_union_linked_data,
+        ),
+    ]